requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
av==12.3.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
ctranslate2==4.3.1
faster-whisper==1.0.3
filelock==3.24.3
Flask==3.0.2
flask-cors==6.0.2
flatbuffers==25.12.19
fsspec==2026.2.0
gTTS==2.5.1
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
mysql-connector-python==8.0.33
numpy==2.4.2
onnxruntime==1.24.2
packaging==26.0
protobuf==6.33.5
Pygments==2.19.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
setuptools==82.0.0
shellingham==1.5.4
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.23.1
typer-slim==0.23.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
